fix(cli): split MCP capability provider to avoid pulling runtime-host/server into TUI
TUI imported createMcpCapabilityProvider from
runtime-host-capability-provider-command.ts, which statically imports
runRuntimeHostProcessLifecycle from @maka/runtime-host/server. That
pulled 372 runtime files and 879 total ESM modules, making
runtime-host-tui-command import ~1.28s vs 0.23s for cli-core.

Split the pure createMcpCapabilityProvider into
mcp-capability-provider.ts (deps only on @maka/core/mcp,
@maka/mcp, @maka/runtime-host/protocol). tui-mcp-control now
imports from the pure module; the command file re-exports for
compatibility and keeps the server import only for
runRuntimeHostCapabilityProviderCli.

Fixes #4005

Measured:
- tui-command import 1280ms/879 files/233 runtime -> 563ms/287 files/3 runtime (-56%)
- tui-mcp-control 1110ms/808 files/233 runtime -> 76 core/5 storage/0 runtime

Generated-by: Muse Spark
